web app/requirements.txt
Flask==1.0.2
gunicorn==20.0.4
googlemaps==4.2.0
mysql-connector==2.2.9
gevent==1.4.0
uwsgi==2.0.18